pyproject.toml
[tool.poetry]
name = "ml-tech-assessment"
version = "0.1.0"
description = ""
authors = ["jhonvalderrama <jhonvalderrama@aceup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.76.2"
pydantic-settings = "^2.9.1"
pytest = "^8.3.5"
fastapi = "^0.129.1"
uvicorn = "^0.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
